refactor: extract AbstractMailBox loop from GoogleMailBase into base/mail.py
